support inlining __torch_function__ with reading from closure
Summary:

The previous PRs to add `__torch_function__` support inlined through
`__torch_function__` without adding any guards for the function.

This worked for simple cases, but did not work if `__torch_function__`
needs to read a nonlocal variable, for example: https://gist.github.com/vkuzo/a3388fcaa532318d049368e96652b366
The reason it was broken is because the code which bound arguments
during inlining had to have a reference to a source in order to bind
things properly.

One way to fix this is to get the source of the `__torch_function__` attribute
of the original tensor, guard on it, and persist it through
all the rewrapping logic.

I'm flexible if there is a better alternative, lmk.

Test plan:

```
pytest -vsk test_torch_function_with_closure
```