Make eval_frame thread safe (#239)
This should make eval_frame thread safe. Currently, the eval_frame is a global object, and different threads my step on each other setting a different one.

This changes the behavior to instead always* have a "shim" eval_frame which then routes to the correct behavior by looking at the thread-local associated object. This is thread safe because now the callback object is always thread safe, and we only use it to drive logic at frame eval time, as opposed to at callback registration time.

Currently, the logic for None/False/Callback is kept, but the False case could be easily collapsed behind the shim in a subsequent diff.

*Always here means always when dynamo is running. The shim is installed and removed based on keeping track of how many dynamo threads are running at the moment.