[Draft] Light Refactor + Add support for torch.autograd.profiler.record_function
This diff takes GradModeVariable's logic and pulls it partially into a more generic ContextWrappingVariable base class intended for making it easier to write context managed code.

This works, and passes all tests, but I am somewhat conflicted on the refactor. On one hand, it pulls a lot of boilerplate out, on the other hand, so far we dont have a ton of cases and it does make it more complicated to read. I also do not know what the n+1 case will look like, in terms of shared vs overriden logic.