Update on "Use fake mode when doing mutation analysis"
Currently, the mutation analysis is done without enabling fake
tensor mode.  This means if there is compute on a constant it
won't convert into fake tensor, and will subsequently fail
in fake tensor mode (which doesn't support mixed fake and non fake
inputs.)  The most likely situation this occurs is when accessing
tensor constants on the GraphModule.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

[ghstack-poisoned]