add check for rematerialization (#644)
realize a node if it's user is output.

If a node's user is output, it must be materialized. Without the change, a node with an output-user might be re-computed by some compute-buffers, but read by other compute-buffers. The recomputation can be saved by forcing the node to be materialized early.

This might make some benchmark slower because of the current fusing strategy sometimes block vertical fusion by doing horizontal fusion too early. This should be solved when we have a more optimal fusing strategy.