Delete python key mode
AOTAutograd subsumes this mode (in particular, if AOTAutograd
gets a graph with no requires grad, or with no_grad on, it should not bother to
wrap the function in a custom autograd.Function); so if you have feature
requests for this, send them to AOTAutograd.  If you need a cheap and
cheerful lowering, you can use make_fx().

Signed-off-by: Edward Z. Yang <ezyang@fb.com>