Improve torchdynamo.explain() explanation (#1218)
Improve the stack trace to point to the actual location of the graph
break.

Before, we were lazily computing the stack trace at the time we were
compiling the subgraph. However, if were graph breaking inside an
inlined function, the sequence of events looked like:
1. Catch the graph break exception
2. Restore the old graph state <-- thus overriding the user stack location
3. Compute the stack trace and add it as a message to the compiled graph

So the graph break explanation pointed to the caller, not the actual
location of the graph break. By computing the stack trace at the same
time we're processing the graph break exception, we can fix that.

This also starts printing the whole stack trace that we have instead of
just the most recent line, which I think is more useful (but if people
find it noisy I can change it back).

Looking for advice on testing.