[export] dedupe break reasons in explain() output
Currently, the `explain()` summary lists the graph break reason for
every graph. This can lead to repetitive output when either:
- A single graph break produces multiple graphs (happens for jumps)
- A single instruction is called many times in the same model and causes
  a graph break.

This PR dedupes the break reasons listed in the `explain()` summary by
the innermost stack frame referenced, which should reduce the verbose of
the explanation.

This is not always what you want--for example, maybe you want to track
down all the different stack traces that lead to a graph break, even if
they end in the same actual instruction. So I also added an additional
output that returns the un-deduped graph break reasons. At some point we
should make a nice API for interacting with these.