Fix #1466 (#1469)
Fix #1466

The general strategy is to represent each Size of a torch.Size individually, rather than also having a variable that collects everything. However, we have to workaround problems with torch.Size not accepting Node/Proxy as members. There's also an optimization to directly use a torch.Size proxy if we have one (which sometimes we do.)

Signed-off-by: Edward Z. Yang <ezyang@fb.com>