DDPOptimizer: recompile graph after replacing subgraphs
DDPOptimizer does the following:
1. split up the graph into submodules
2. replace each of the submodules with a compiled submodule

After step 2, we weren't calling `graph_module.recompile()`, which
caused us to call the old submodules. The old submodules didn't have the
step that unwraps tuples, which caused type errors, described below.

The old submodules will wrap outputs in tuples, even if there's only a
single output. So we might have submodule1 which should output a tensor
that is consumed by submodule2. In this case we need to unwrap the
output tuple from submodule 1; that's what WrapperModule does. But since
we weren't recompiling, the graph_module was still calling the old
submodules.

For some reason, this issue only occurred when using inductor backend.